Reorganize and rename "Test Vehicles" to "Prebuilt Vehicles"
